Refactored UpdateSeries out of Scanner and into a dedicated file.
